chore: schedule cron job to refresh token
