+++ src/PaymentGateways/PayPalCommerce/PayPalClient.php
@@ -32,19 +32,6 @@ class PayPalClient {
 	}
 
 	/**
-	 * Bootstrap class.
-	 *
-	 * @since 2.8.0
-	 *
-	 * @return PayPalClient
-	 */
-	public function boot() {
-		add_action( 'give_paypal_commerce_refresh_token', [ $this, 'refreshToken' ], 10, 2 );
-
-		return $this;
-	}
-
-	/**
 	 * Get environment.
 	 *
 	 * @sicne 2.8.0
@@ -69,50 +56,5 @@ class PayPalClient {
 	 */
 	public function getHttpClient() {
 		return new PayPalHttpClient( $this->getEnvironment() );
-	}
-
-	/**
-	 * Register cron job to refresh access token.
-	 * Note: only for internal use.
-	 *
-	 * @param  string  $tokenExpires
-	 *
-	 * @since 2.8.0
-	 *
-	 */
-	public function registerCronJobTorRefreshToken( $tokenExpires ) {
-		wp_schedule_single_event(
-			time() + ( $tokenExpires - 1800 ), // Refresh token before half hours of expires date.
-			'give_paypal_commerce_refresh_token'
-		);
-	}
-
-	/**
-	 * Refresh token.
-	 * Note: only for internal use
-	 *
-	 * @since 2.8.0
-	 */
-	public function refreshToken() {
-		/* @var MerchantDetail $merchant */
-		$merchant = give( MerchantDetail::class );
-
-		$refreshToken  = $merchant->getRefreshToken();
-		$request       = new RefreshTokenRequest(
-			$this->getEnvironment(),
-			$refreshToken
-		);
-		$request->body = [
-			'grant_type'    => 'refresh_token',
-			'refresh_token' => $refreshToken,
-		];
-
-		$response     = $this->getHttpClient()->execute( $request );
-		$tokenDetails = ArrayDataSet::ucWordInKeyNameComesAfterDash( $response->result );
-
-		$merchant->setTokenDetails( $tokenDetails );
-		$merchant->save();
-
-		$this->registerCronJobTorRefreshToken( $tokenDetails['expiresIn'] );
 	}
 }
