+++ src/PaymentGateways/PayPalCommerce/PayPalCommerce.php
@@ -70,11 +70,18 @@ class PayPalCommerce implements PaymentGateway {
 			}
 		);
 
+		give()->singleton( PayPalClient::class );
+
 		give()->singleton(
-			PayPalClient::class,
-			static function () {
-				return ( new PayPalClient() )->boot();
+			RefreshToken::class,
+			static function() {
+				return ( new RefreshToken() )->boot();
 			}
 		);
+
+		// Boot RefreshToken class immediately if cron job running.
+		if ( wp_doing_cron() ) {
+			give( RefreshToken::class );
+		}
 	}
 }
